requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
banks==2.1.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emoji==2.14.1
eval_type_backport==0.2.2
faiss-cpu==1.10.0
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langdetect==1.0.9
langsmith==0.3.23
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.28
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.28
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.10
llama-index-llms-openai==0.3.29
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
olefile==0.47
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pymongo==4.11.3
PyMySQL==1.1.1
pypdf==5.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.32.0
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
